+++ lib/client.js
@@ -2,7 +2,7 @@
 
 const events = require("events");
 const util = require("util");
-const { throwNotSupported } = require("./new_utils.js");
+const { throwNotSupported } = require("./new-utils.js");
 
 const utils = require("./utils.js");
 const errors = require("./errors.js");
