+++ lib/geometry/line-string.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { throwNotSupported } = require("../new_utils");
+const { throwNotSupported } = require("../new-utils");
 
 /**
  * @deprecated Not supported by the driver. Usage will throw an error.
